reorganize swmm_enum and toolkit_enum into solver_enum
